osusec/relevant_xkcd/relevantxkcd/requirements.txt
discord.py==2.2.3
openai==0.27.2
